+++ src/grepros/__init__.py
@@ -11,8 +11,8 @@ Released under the BSD License.
 """
 ## @namespace grepros
 __title__        = "grepros"
-__version__      = "1.2.0.dev0"
-__version_info__ = (1, 2, 0, "dev0")
+__version__      = "1.2.0.dev1"
+__version_info__ = (1, 2, 0, "dev1")
 __version_date__ = "13.01.2024"
 
 from . library import *